// src/types/api.ts
export interface NobelPrize {
  /** @example 1989 */
  awardYear?: number;
  category?: Translation;
  categoryFullName?: Translation;
  /**
   * @format date
   * @example "2015-11-12"
   */
  dateAwarded?: string;
  /**
   * Prize amount given at the year of the award
   * @example 1000000
   */
  prizeAmount?: number;
  /**
   * Prize amount adjusted according to Index number yearly average
   * @example 1000000
   */
  prizeAmountAdjusted?: number;
  topMotivation?: Translation;
  laureates?: LaureateBasic[];
}
// ...
export interface LaureateBasic {
  /**
   * @min 1
   * @example 34
   */
  id?: number;
  name?: Translation;
  /** @default "1" */
  portion?: "1" | "1/2" | "1/3" | "1/4";
  /** @default "1" */
  sortOrder?: "1" | "2" | "3";
  motivation?: Translation;
  links?: ItemLinks[];
}
// ...
export interface ItemLinks {
  rel?: string;
  href?: string;
  action?: string;
  types?: string;
}
// ...
export interface Translation {
  /** @example "English" */
  en?: string;
  /** @example "Svenska" */
  se?: string;
  /** @example "Norsk" */
  false?: string;
}